feat: load editPost or createPost on state change
