+++ frontend/src/pages/Home.js
@@ -1,12 +1,15 @@
-import React from "react";
+import React, { useState } from "react";
 import Thread from "../components/Thread";
 import PostForm from "../components/Post/PostForm";
 
 const Home = () => {
+
+    const [createPost, setCreatePost] = useState({})
+
     return (
         <main className='home'>
-            <PostForm />
-            <Thread />
+            <PostForm setCreatePost={setCreatePost}/>
+            <Thread createPost={createPost}/>
         </main>
     );
 };
